refactor: rename plugin to kebab-case
