+++ packages/entities/entities-plugins/src/components/free-form/plugins/ai-mcp-proxy/index.ts
