+++ packages/entities/entities-plugins/src/components/free-form/plugins/basic-auth.ts
