+++ packages/entities/entities-plugins/src/components/free-form/plugins/cors.ts
