+++ packages/entities/entities-plugins/src/components/free-form/plugins/exit-transformer.ts
