+++ packages/entities/entities-plugins/src/components/free-form/plugins/key-auth/KeyAuthForm.vue
