+++ packages/entities/entities-plugins/src/components/free-form/plugins/service-protection/ConfigForm.vue
